+++ apps/android/android/app/src/main/java/de/correlcore/app/MainActivity.java
@@ -1,5 +1,17 @@
 package de.correlcore.app;
 
+import android.os.Bundle;
+
 import com.getcapacitor.BridgeActivity;
 
-public class MainActivity extends BridgeActivity {}
+import de.correlcore.app.widget.WidgetCredentialsPlugin;
+import de.correlcore.app.widget.WidgetRefreshWorker;
+
+public class MainActivity extends BridgeActivity {
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        registerPlugin(WidgetCredentialsPlugin.class);
+        super.onCreate(savedInstanceState);
+        WidgetRefreshWorker.enqueuePeriodic(getApplicationContext());
+    }
+}